+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Python Versioning](http://legacy.python.org/dev/pep
 ## [Unreleased]
 * #### ALL
   * #### Added
+    * Support for ViInt64 (64-bit integers)
   * #### Changed
     * Modified how methods with repeated capabilities are invoked. There's no longer (for example) a "channel_name" input. Instead:
       ```python